surveillance-system/backend/requirements.txt
# FastAPI Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Object Storage
minio==7.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
httpx==0.26.0
aiofiles==23.2.1
orjson==3.9.10

# Image Processing
pillow==10.2.0
numpy==1.26.3

# WebSockets
websockets==12.0
socketio==0.2.1
python-socketio==5.11.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
